third_party/WebKit/LayoutTests/http/tests/fetch/chromium/mime-sniffing.html

<!doctype html>
<script src = "/resources/testharness.js"></script>
<script src = "/resources/testharnessreport.js"></script>
<script>
promise_test(() => {
  return fetch('/fetch/chromium/resources/mime-sniffing.php').then(res => {
      return res.blob();
    }).then(blob => {
      assert_equals(blob.type, 'unknown/unknown');
    });
  }, 'Mime sniffing should not run.');
</script>
